requirements.txt
# requirements.txt

blinker==1.7.0
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
colormap==1.0.6
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
easydev==0.13.1
Flask==3.0.3
fonttools==4.51.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
line-profiler==4.1.2
MarkupSafe==2.1.5
matplotlib==3.8.4
nodeenv==1.8.0
numpy==1.26.4
packaging==24.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
ptyprocess==0.7.0
pyparsing==3.1.2
pypng==0.20220715.0
pyright==1.1.360
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
python-dateutil==2.9.0.post0
qrcode==7.4.2
ruff==0.4.2
shiboken6==6.7.0
six==1.16.0
typing_extensions==4.11.0
Werkzeug==3.0.2

